+++ temba/channels/types/telesom/type.py
@@ -36,9 +36,7 @@ class TelesomType(ChannelType):
         dict(label="", url="https://{{ channel.callback_domain }}{% url 'courier.ts' channel.uuid 'receive' %}"),
     )
 
-    def is_available_to(self, user):
-        org = user.get_org()
-        return org.timezone and str(org.timezone) in ["Africa/Mogadishu"]
+    available_timezones = ["Africa/Mogadishu"]
 
     def is_recommended_to(self, user):
-        return self.is_available_to(user)
+        return self.is_available_to(user)[0]