+++ temba/orgs/tests.py
@@ -11,6 +11,7 @@ import stripe
 import stripe.error
 from bs4 import BeautifulSoup
 from dateutil.relativedelta import relativedelta
+from smartmin.users.models import FailedLogin
 
 from django.conf import settings
 from django.contrib.auth.models import Group, User
@@ -185,6 +186,77 @@ class UserTest(TembaTest):
 
         self.assertEqual(9, len(self.admin.backup_tokens.filter(is_used=False)))
 
+    @override_settings(USER_LOCKOUT_TIMEOUT=1, USER_FAILED_LOGIN_LIMIT=3)
+    def test_login_lockouts(self):
+        login_url = reverse("users.user_login")
+        verify_url = reverse("users.two_factor_verify")
+        backup_url = reverse("users.two_factor_backup")
+        failed_url = reverse("users.user_failed")
+
+        # submit incorrect username and password 3 times
+        self.client.post(login_url, {"username": "Administrator", "password": "pass123"})
+        self.client.post(login_url, {"username": "Administrator", "password": "pass123"})
+        response = self.client.post(login_url, {"username": "Administrator", "password": "pass123"})
+
+        self.assertRedirect(response, failed_url)
+        self.assertRedirect(self.client.get(reverse("msgs.msg_inbox")), login_url)
+
+        # simulate failed logins timing out by making them older
+        FailedLogin.objects.all().update(failed_on=timezone.now() - timedelta(minutes=3))
+
+        # now we're allowed to make failed logins again
+        response = self.client.post(login_url, {"username": "Administrator", "password": "pass123"})
+        self.assertFormError(
+            response,
+            "form",
+            "__all__",
+            "Please enter a correct username and password. Note that both fields may be case-sensitive.",
+        )
+
+        # and successful logins
+        response = self.client.post(login_url, {"username": "Administrator", "password": "Administrator"})
+        self.assertRedirect(response, reverse("orgs.org_choose"))
+
+        # try again with 2FA enabled
+        self.client.logout()
+        self.admin.enable_2fa()
+
+        # submit incorrect username and password 3 times
+        self.client.post(login_url, {"username": "Administrator", "password": "pass123"})
+        self.client.post(login_url, {"username": "Administrator", "password": "pass123"})
+        response = self.client.post(login_url, {"username": "Administrator", "password": "pass123"})
+
+        self.assertRedirect(response, failed_url)
+        self.assertRedirect(self.client.get(reverse("msgs.msg_inbox")), login_url)
+
+        # login correctly
+        FailedLogin.objects.all().delete()
+        response = self.client.post(login_url, {"username": "Administrator", "password": "Administrator"})
+        self.assertRedirect(response, verify_url)
+
+        # now enter a backup token 3 times incorrectly
+        self.client.post(backup_url, {"token": "nope"})
+        self.client.post(backup_url, {"token": "nope"})
+        response = self.client.post(backup_url, {"token": "nope"})
+
+        self.assertRedirect(response, failed_url)
+        self.assertRedirect(self.client.get(verify_url), login_url)
+        self.assertRedirect(self.client.get(backup_url), login_url)
+        self.assertRedirect(self.client.get(reverse("msgs.msg_inbox")), login_url)
+
+        # simulate failed logins timing out by making them older
+        FailedLogin.objects.all().update(failed_on=timezone.now() - timedelta(minutes=3))
+
+        # we can't enter backup tokens again without going thru regular login first
+        response = self.client.post(backup_url, {"token": "nope"})
+        self.assertRedirect(response, login_url)
+
+        response = self.client.post(login_url, {"username": "Administrator", "password": "Administrator"})
+        self.assertRedirect(response, verify_url)
+
+        response = self.client.post(backup_url, {"token": self.admin.backup_tokens.first()})
+        self.assertRedirect(response, reverse("orgs.org_choose"))
+
     def test_two_factor(self):
         self.assertFalse(self.admin.get_settings().two_factor_enabled)
 
@@ -220,9 +292,6 @@ class UserTest(TembaTest):
         self.assertFalse(self.admin.get_settings().two_factor_enabled)
 
     def test_two_factor_views(self):
-        # 2FA still only accessible to beta users
-        Group.objects.get(name="Beta").user_set.add(self.admin)
-
         enable_url = reverse("orgs.user_two_factor_enable")
         tokens_url = reverse("orgs.user_two_factor_tokens")
         disable_url = reverse("orgs.user_two_factor_disable")
@@ -290,6 +359,28 @@ class UserTest(TembaTest):
         # trying to view the tokens page now takes us to the enable form
         response = self.client.get(tokens_url)
         self.assertRedirect(response, enable_url)
+
+    def test_two_factor_time_limit(self):
+        login_url = reverse("users.user_login")
+        verify_url = reverse("users.two_factor_verify")
+        backup_url = reverse("users.two_factor_backup")
+
+        self.admin.enable_2fa()
+
+        # simulate a login for a 2FA user 10 minutes ago
+        with patch("django.utils.timezone.now", return_value=timezone.now() - timedelta(minutes=10)):
+            response = self.client.post(login_url, {"username": "Administrator", "password": "Administrator"})
+            self.assertRedirect(response, verify_url)
+
+            response = self.client.get(verify_url)
+            self.assertEqual(200, response.status_code)
+
+        # if they access the verify or backup page now, they are redirected back to the login page
+        response = self.client.get(verify_url)
+        self.assertRedirect(response, login_url)
+
+        response = self.client.get(backup_url)
+        self.assertRedirect(response, login_url)
 
     def test_ui_permissions(self):
         # non-logged in users can't go here
@@ -1343,8 +1434,6 @@ class OrgTest(TembaTest):
         self.assertEqual("norkans7@gmail.com", invitation.email)
         self.assertEqual("A", invitation.user_group)
 
-        old_secret = invitation.secret
-
         # and sent by email
         self.assertEqual(1, len(mail.outbox))
 
@@ -1355,28 +1444,6 @@ class OrgTest(TembaTest):
         # no longer appears in list
         response = self.client.get(url)
         self.assertNotContains(response, "norkans7@gmail.com")
-
-        # send another invitation, different role
-        self.client.post(
-            url,
-            {
-                f"user_{self.admin.id}_role": "A",
-                f"user_{self.editor.id}_role": "E",
-                f"user_{self.user.id}_role": "E",
-                f"user_{self.agent.id}_role": "T",
-                "invite_emails": "norkans7@gmail.com",
-                "invite_role": "E",
-            },
-        )
-
-        # old invite should be updated with new secret
-        invitation.refresh_from_db()
-        self.assertEqual("E", invitation.user_group)
-        self.assertTrue(invitation.is_active)
-        self.assertNotEqual(old_secret, invitation.secret)
-
-        # and new email sent
-        self.assertEqual(2, len(mail.outbox))
 
         # include multiple emails on the form
         self.client.post(
@@ -1393,7 +1460,7 @@ class OrgTest(TembaTest):
 
         # now 2 new invitations are created and sent
         self.assertEqual(3, Invitation.objects.all().count())
-        self.assertEqual(4, len(mail.outbox))
+        self.assertEqual(3, len(mail.outbox))
 
         response = self.client.get(url)
 
@@ -1403,7 +1470,7 @@ class OrgTest(TembaTest):
 
         # invites ordered by email as well
         invites_on_form = [row["invite"].email for row in response.context["form"].invite_rows]
-        self.assertEqual(["code@temba.com", "norbert@temba.com", "norkans7@gmail.com"], invites_on_form)
+        self.assertEqual(["code@temba.com", "norbert@temba.com"], invites_on_form)
 
         # try to remove ourselves as admin
         response = self.client.post(
@@ -1471,18 +1538,102 @@ class OrgTest(TembaTest):
         self.admin.refresh_from_db()
         self.assertEqual(self.admin.api_tokens.filter(is_active=True).count(), 0)
 
+        # make sure an existing user can not be invited again
+        user = Org.create_user("admin1@temba.com", "admin1@temba.com")
+        user.set_org(self.org)
+        self.org.administrators.add(user)
+        self.login(user)
+
+        self.assertEqual(1, Invitation.objects.filter(is_active=True).count())
+
+        # include multiple emails on the form
+        response = self.client.post(
+            url,
+            {
+                f"user_{self.admin.id}_role": "A",
+                f"user_{self.editor.id}_role": "E",
+                f"user_{self.user.id}_role": "E",
+                f"user_{self.agent.id}_role": "T",
+                f"user_{user.id}_role": "A",
+                "invite_emails": "norbert@temba.com,code@temba.com,admin1@temba.com",
+                "invite_role": "A",
+            },
+        )
+
+        self.assertFormError(
+            response, "form", "invite_emails", "One of the emails you entered has an existing user on the workspace."
+        )
+
+        # do not allow multiple invite on the same email
+        response = self.client.post(
+            url,
+            {
+                f"user_{self.admin.id}_role": "A",
+                f"user_{self.editor.id}_role": "E",
+                f"user_{self.user.id}_role": "E",
+                f"user_{self.agent.id}_role": "T",
+                f"user_{user.id}_role": "A",
+                "invite_emails": "norbert@temba.com,code@temba.com",
+                "invite_role": "A",
+            },
+        )
+
+        self.assertFormError(
+            response, "form", "invite_emails", "One of the emails you entered has an existing user on the workspace."
+        )
+
+        # no error for inactive invite
+        response = self.client.post(
+            url,
+            {
+                f"user_{self.admin.id}_role": "A",
+                f"user_{self.editor.id}_role": "E",
+                f"user_{self.user.id}_role": "E",
+                f"user_{self.agent.id}_role": "T",
+                f"user_{user.id}_role": "A",
+                "invite_emails": "code@temba.com, code@temba.com",
+                "invite_role": "A",
+            },
+        )
+
+        self.assertFormError(response, "form", "invite_emails", "One of the emails you entered is duplicated.")
+
+        # no error for inactive invite
+        response = self.client.post(
+            url,
+            {
+                f"user_{self.admin.id}_role": "A",
+                f"user_{self.editor.id}_role": "E",
+                f"user_{self.user.id}_role": "E",
+                f"user_{self.agent.id}_role": "T",
+                f"user_{user.id}_role": "A",
+                "invite_emails": "code@temba.com",
+                "invite_role": "A",
+            },
+        )
+
+        self.assertEqual(2, Invitation.objects.filter(is_active=True).count())
+        self.assertTrue(Invitation.objects.filter(is_active=True, email="code@temba.com").exists())
+        self.assertEqual(4, len(mail.outbox))
+
     @patch("temba.utils.email.send_temba_email")
     def test_join(self, mock_send_temba_email):
-        def create_invite(group):
+        def create_invite(group, username):
             return Invitation.objects.create(
                 org=self.org,
                 user_group=group,
-                email="norkans7@gmail.com",
+                email=f"{username}@nyaruka.com",
                 created_by=self.admin,
                 modified_by=self.admin,
             )
 
-        editor_invitation = create_invite("E")
+        def create_user(username):
+            user = User.objects.create_user(f"{username}@nyaruka.com", f"{username}@nyaruka.com")
+            user.set_password(f"{username}@nyaruka.com")
+            user.save()
+            return user
+
+        editor_invitation = create_invite("E", "invitededitor")
         editor_invitation.send()
         email_args = mock_send_temba_email.call_args[0]  # all positional args
 
@@ -1504,10 +1655,35 @@ class OrgTest(TembaTest):
         )
 
         # a user is already logged in
-        self.invited_editor = self.create_user("InvitedEditor")
-        self.login(self.invited_editor)
+        self.invited_editor = create_user("invitededitor")
+
+        # different user login
+        self.login(self.admin)
 
         response = self.client.get(editor_join_url)
+        self.assertEqual(200, response.status_code)
+
+        # should be logged out to request login
+        self.assertEqual(0, len(self.client.session.keys()))
+
+        # login with a diffent user that the invited
+        self.login(self.admin)
+        response = self.client.get(reverse("orgs.org_join_accept", args=[editor_invitation.secret]), follow=True)
+        self.assertEqual(200, response.status_code)
+        self.assertEqual(response.request["PATH_INFO"], reverse("orgs.org_join", args=[editor_invitation.secret]))
+
+        self.login(self.invited_editor)
+        response = self.client.get(editor_join_url)
+        self.assertEqual(302, response.status_code)
+        response = self.client.get(editor_join_url, follow=True)
+        self.assertEqual(
+            response.request["PATH_INFO"], reverse("orgs.org_join_accept", args=[editor_invitation.secret])
+        )
+
+        editor_join_accept_url = reverse("orgs.org_join_accept", args=[editor_invitation.secret])
+        self.login(self.invited_editor)
+
+        response = self.client.get(editor_join_accept_url)
         self.assertEqual(200, response.status_code)
 
         self.assertEqual(self.org.pk, response.context["org"].pk)
@@ -1515,7 +1691,7 @@ class OrgTest(TembaTest):
         self.assertEqual(1, len(response.context["form"].fields))
 
         post_data = dict()
-        response = self.client.post(editor_join_url, post_data, follow=True)
+        response = self.client.post(editor_join_accept_url, post_data, follow=True)
         self.assertEqual(200, response.status_code)
 
         self.assertIn(self.invited_editor, self.org.editors.all())
@@ -1530,22 +1706,29 @@ class OrgTest(TembaTest):
 
         # test it for each role
         for role in roles:
-            invite = create_invite(role[0])
-            user = self.create_user("User%s" % role[0])
+            invite = create_invite(role[0], "User%s" % role[0])
+            user = create_user("User%s" % role[0])
             self.login(user)
-            response = self.client.post(reverse("orgs.org_join", args=[invite.secret]), follow=True)
+            response = self.client.post(reverse("orgs.org_join_accept", args=[invite.secret]), follow=True)
             self.assertEqual(200, response.status_code)
             self.assertIsNotNone(role[1].filter(pk=user.pk).first())
 
         # try an expired invite
-        invite = create_invite("S")
+        invite = create_invite("S", "invitedexpired")
         invite.is_active = False
         invite.save()
-        expired_user = self.create_user("InvitedExpired")
+        expired_user = create_user("invitedexpired")
         self.login(expired_user)
-        response = self.client.post(reverse("orgs.org_join", args=[invite.secret]), follow=True)
+        response = self.client.post(reverse("orgs.org_join_accept", args=[invite.secret]), follow=True)
         self.assertEqual(200, response.status_code)
         self.assertIsNone(self.org.surveyors.filter(pk=expired_user.pk).first())
+
+        response = self.client.post(reverse("orgs.org_join", args=[invite.secret]))
+        self.assertEqual(302, response.status_code)
+
+        response = self.client.post(reverse("orgs.org_join", args=[invite.secret]), follow=True)
+        self.assertEqual(200, response.status_code)
+        self.assertEqual(response.request["PATH_INFO"], reverse("users.user_login"))
 
     def test_create_login(self):
         admin_invitation = Invitation.objects.create(
@@ -1560,17 +1743,15 @@ class OrgTest(TembaTest):
 
         self.assertEqual(self.org.pk, response.context["org"].pk)
 
-        # we have a form with 4 fields and one hidden 'loc'
-        self.assertEqual(5, len(response.context["form"].fields))
+        # we have a form with 3 fields and one hidden 'loc'
+        self.assertEqual(4, len(response.context["form"].fields))
         self.assertIn("first_name", response.context["form"].fields)
         self.assertIn("last_name", response.context["form"].fields)
-        self.assertIn("email", response.context["form"].fields)
         self.assertIn("password", response.context["form"].fields)
 
         post_data = dict()
         post_data["first_name"] = "Norbert"
         post_data["last_name"] = "Kwizera"
-        post_data["email"] = "norkans7@gmail.com"
         post_data["password"] = "norbertkwizeranorbert"
 
         response = self.client.post(admin_create_login_url, post_data, follow=True)
@@ -1580,6 +1761,28 @@ class OrgTest(TembaTest):
         self.assertTrue(new_invited_user in self.org.administrators.all())
         self.assertFalse(Invitation.objects.get(pk=admin_invitation.pk).is_active)
 
+        invitation = Invitation.objects.create(
+            org=self.org, user_group="E", email="norkans7@gmail.com", created_by=self.admin, modified_by=self.admin
+        )
+        create_login_url = reverse("orgs.org_create_login", args=[invitation.secret])
+
+        # we have a matching user so we redirect with the user logged in
+        response = self.client.get(create_login_url)
+        self.assertEqual(302, response.status_code)
+
+        response = self.client.get(create_login_url, follow=True)
+        self.assertEqual(200, response.status_code)
+        self.assertEqual(response.request["PATH_INFO"], reverse("orgs.org_join_accept", args=[invitation.secret]))
+
+        invitation.is_active = False
+        invitation.save()
+
+        response = self.client.get(create_login_url)
+        self.assertEqual(302, response.status_code)
+        response = self.client.get(create_login_url, follow=True)
+        self.assertEqual(200, response.status_code)
+        self.assertEqual(response.request["PATH_INFO"], reverse("public.public_index"))
+
     def test_create_login_invalid_form(self):
         admin_invitation = Invitation.objects.create(
             org=self.org, user_group="A", email="user@example.com", created_by=self.admin, modified_by=self.admin
@@ -1588,24 +1791,9 @@ class OrgTest(TembaTest):
         admin_create_login_url = reverse("orgs.org_create_login", args=[admin_invitation.secret])
         self.client.logout()
 
-        post_data = dict(first_name="Matija", last_name="Vujica", email="", password="just-a-password")
-
-        response = self.client.post(admin_create_login_url, post_data, follow=True)
-
-        self.assertFormError(response, "form", "email", "This field is required.")
-
-        post_data = dict(
-            first_name="Matija", last_name="Vujica", email="this-is-not-a-valid-email", password="just-a-password"
-        )
-
-        response = self.client.post(admin_create_login_url, post_data, follow=True)
-
-        self.assertFormError(response, "form", "email", "Enter a valid email address.")
-
         post_data = dict(
             first_name="Matija_first_name_longer_than_30_chars",
             last_name="Vujica_last_name_longer_than_150_chars____lorem-ipsum-dolor-sit-amet-ipsum-dolor-sit-amet-ipsum-dolor-sit-amet-ipsum-dolor-sit-amet-ipsum-dolor-sit-amet-ipsum-dolor-sit-amet",
-            email="matija@vujica-this-is-a-verrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrry-loooooooooooooooooooooooooong-domain-name-not-sure-if-this-is-even-possible-to-register.com",
             password="just-a-password",
         )
         response = self.client.post(admin_create_login_url, post_data)
@@ -1615,8 +1803,6 @@ class OrgTest(TembaTest):
         self.assertFormError(
             response, "form", "last_name", "Ensure this value has at most 150 characters (it has 173)."
         )
-        self.assertFormError(response, "form", "email", "Ensure this value has at most 150 characters (it has 161).")
-        self.assertFormError(response, "form", "email", "Enter a valid email address.")
 
     def test_surveyor_invite(self):
         surveyor_invite = Invitation.objects.create(
@@ -4956,14 +5142,13 @@ class EmailContextProcessorsTest(TembaTest):
         with self.settings(HOSTNAME="rapidpro.io"):
             forget_url = reverse("users.user_forget")
 
-            post_data = dict()
-            post_data["email"] = "nouser@nouser.com"
-
-            self.client.post(forget_url, post_data, follow=True)
+            response = self.client.post(forget_url, {"email": "Administrator@nyaruka.com"})
+            self.assertLoginRedirect(response)
             self.assertEqual(1, len(mail.outbox))
+
             sent_email = mail.outbox[0]
             self.assertEqual(len(sent_email.to), 1)
-            self.assertEqual(sent_email.to[0], "nouser@nouser.com")
+            self.assertEqual(sent_email.to[0], "Administrator@nyaruka.com")
 
             # we have the domain of rapipro.io brand
             self.assertIn("app.rapidpro.io", sent_email.body)